+++ lib/Input/Request.php
@@ -10,7 +10,7 @@
 namespace SebLucas\Cops\Input;
 
 use SebLucas\Cops\Calibre\Filter;
-use SebLucas\Cops\Output\Format;
+use SebLucas\Cops\Output\Response;
 
 /**
  * Summary of Request
@@ -124,7 +124,7 @@ class Request
             $params = Route::match($path);
             if (is_null($params)) {
                 // this will call exit()
-                $this->notFound(null, "Invalid request path '$path'");
+                Response::sendError($this, "Invalid request path '$path'");
             }
             // @todo handle 'json' routes correctly - see util.js
             if (empty($params[Route::HANDLER_PARAM]) && $this->isAjax()) {
@@ -553,32 +553,6 @@ class Request
             return true;
         }
         return false;
-    }
-
-    /**
-     * Summary of notFound
-     * @param string|null $home
-     * @param string|null $error
-     * @param array<string, mixed> $params
-     * @return never
-     */
-    public static function notFound($home = null, $error = null, $params = [])
-    {
-        header(($_SERVER['SERVER_PROTOCOL'] ?? 'HTTP/1.1') . ' 404 Not Found');
-        header('Status: 404 Not Found');
-
-        $_SERVER['REDIRECT_STATUS'] = 404;
-        // default to script basename or null if undefined
-        $home ??= basename($_SERVER['SCRIPT_NAME'] ?? '') ?: null;
-        $data = ['link' => Route::url($home, null, $params)];
-        if (!empty($error)) {
-            $data['error'] = htmlspecialchars($error);
-            $template = 'templates/error.html';
-        } else {
-            $template = 'templates/notfound.html';
-        }
-        echo Format::template($data, $template);
-        exit;
     }
 
     /**